// src/eNum.tsx
export enum key {
    AC = 'AC',
    C = 'C',
    Del = 'DEL',
    plus = '+',
    minus = '-',
    multiply = '*',
    divide = '/',
    equal = '=',
    point = '.',
    negative = '-',
    num1 = 1,
    num2 = 2,
    num3 = 3,
    num4 = 4,
    num5 = 5,
    num6 = 6,
    num7 = 7,
    num8 = 8,
    num9 = 9,
    num0 = 0,
}